pyproject.toml
[tool.poetry]
name = "spice-draw"
version = "0.1.0"
description = "Generate drawing of schematic from an annotated spice file (based on schemdraw and Jinja2)."
authors = ["Patarimi <38954040+Patarimi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.scripts]
spice2draw = "spice_draw:main.cli"

[tool.poe.tasks]
doc         = "typer spice_draw.main utils docs --name spice2draw --output README.md"
show-update = "poetry show -l -o"
lock-req    = "poetry export -o requirements.txt --without-hashes"
update4commit = ["doc", "lock-req"]

[tool.poetry.dependencies]
python = "^3.8"
schemdraw = "^0.15"
typer = "^0.7.0"
networkx = "^3.0"
matplotlib = "^3.7.1"
pixie-python = "^4.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
typer-cli = "^0.0.13"
poethepoet = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
